DRBG revival: merged drbg_ctr.c into drbg_rand.c
