+++ util/libcrypto.num
@@ -4300,6 +4300,7 @@ ERR_load_strings_const                  4242	1_1_1	EXIST::FUNCTION:
 ASN1_TIME_to_tm                         4243	1_1_1	EXIST::FUNCTION:
 ASN1_TIME_set_string_X509               4244	1_1_1	EXIST::FUNCTION:
 OCSP_resp_get1_id                       4245	1_1_1	EXIST::FUNCTION:OCSP
+<<<<<<< 8530039a307f7aa8acb0516fdd38191baa91d434
 OSSL_STORE_register_loader              4246	1_1_1	EXIST::FUNCTION:
 OSSL_STORE_LOADER_set_error             4247	1_1_1	EXIST::FUNCTION:
 OSSL_STORE_INFO_get0_PKEY               4248	1_1_1	EXIST::FUNCTION:
@@ -4342,6 +4343,40 @@ OSSL_STORE_INFO_set0_NAME_description   4284	1_1_1	EXIST::FUNCTION:
 OSSL_STORE_INFO_get1_NAME_description   4285	1_1_1	EXIST::FUNCTION:
 OSSL_STORE_do_all_loaders               4286	1_1_1	EXIST::FUNCTION:
 OSSL_STORE_LOADER_get0_engine           4287	1_1_1	EXIST::FUNCTION:
+<<<<<<< 8423de55c026b6ec56c736e8c2384ad7cb569ff4
 OPENSSL_fork_prepare                    4288	1_1_1	EXIST:UNIX:FUNCTION:
 OPENSSL_fork_parent                     4289	1_1_1	EXIST:UNIX:FUNCTION:
 OPENSSL_fork_child                      4290	1_1_1	EXIST:UNIX:FUNCTION:
+RAND_DRBG_new                           4291	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_uninstantiate                 4292	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_set_rand_callbacks            4293	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_get_blocklength               4294	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_set_app_data                  4295	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_set_callbacks                 4296	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_free                          4297	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_reseed                        4298	1_1_1	EXIST::FUNCTION:
+RAND_drbg                               4299	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_generate                      4300	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_set_check_interval            4301	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_get_default                   4302	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_instantiate                   4303	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_set_reseed_interval           4304	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_get_app_data                  4305	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_get_strength                  4306	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_new                           4307	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_uninstantiate                 4308	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_set_rand_callbacks            4309	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_get_blocklength               4310	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_set_app_data                  4311	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_set_callbacks                 4312	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_free                          4313	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_reseed                        4314	1_1_1	EXIST::FUNCTION:
+RAND_drbg                               4315	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_generate                      4316	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_set_check_interval            4317	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_get_default                   4318	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_instantiate                   4319	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_set_reseed_interval           4320	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_get_app_data                  4321	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_get_strength                  4322	1_1_1	EXIST::FUNCTION:
+RAND_DRBG_set                           4323	1_1_1	EXIST::FUNCTION:
